more towards the stats of color in the android statues .  fixed a bug in a MiscMath stats method.

+++ src/algorithms/misc/MiscMath.java
@@ -611,7 +611,7 @@ public class MiscMath {
         
         int length = x.size();
         
-        long sumX = 0;
+        double sumX = 0;
         for (int i = 0; i < length; i++) {
             sumX += x.get(i);
         }
